QAPR/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.1
tqdm>=4.62.0

# Information Retrieval
pyserini>=0.20.0
rank-bm25>=0.2.2

# Neural models
torch>=1.10.0
sentence-transformers>=2.2.0
transformers>=4.20.0

# Machine Learning
scikit-learn>=1.0.0
lightgbm>=3.3.0

# Evaluation
pytrec-eval>=0.5

# Text processing
beautifulsoup4>=4.9.3
lxml>=4.6.3
